Refactor AthenaResultSetTest with MockitoExtension and more
MockitoExtension makes it possible to use @Mock to cretate mocks, but their stubbing is non-lenient and will raise if the stubbed method is not used in the tests. To make this work properly I went on a tangent and refactored the way that rows are added to the result set, because I couldn't get the stubbing to work. In the end it turns out that using #reset and #lenient is called for, because we want to be able to override the result rows in inner scopes (which requires #reset, otherwise the next call to #when defines what will be returned on the second call), and not all methods in each scope will call the stubbed method (because they are supposed to throw or whatever, and this means #lenient is a must anyway).