Extract the result loading from AthenaResultSet to helpers
And implement the previous mechanism as StandardResult, and a new slightly better strategy in PreloadingStandardResult. It still uses GetQueryResults, but it keeps a "background" request running to preload the next page while the client is processing the current.

Some modifications to the AthenaResultSet tests were necessary since the pages are not necessarily loaded at the expected points in time anymore.