Add tests for StandardResult
It was covered more or less by AthenaResultSet, but there were some details that were not covered, like exceptions thrown when handling the async result. Most of the tests could be ported easily thanks to GetQueryResultsHelper. To be able to test timeouts and other scenarios where the async loading goes wrong, some additional features were added to GetQueryResultsHelper.

It was really difficult to get CompletableFuture#get raise InterruptedException. In the end I had to resort to Mockito to fake it.