Move #getQueryExecutionId from AthenaResultSet to AthenaResultSetMetaData
It makes more sense to have on the meta data object in a way, and I'm thinking of exposing more of the QueryExecution information, which definitely better belongs there.

This refactoring also makes the QueryExecution information available to Result implementations, which will become important if one of them wants to get hold of the output location, for example.