Read the .metadata file from S3 instead of calling GetQueryResults
The file protobuf encoded, and I've implemented a very basic parser that handles only what's necessary to parse this file. See https://stackoverflow.com/questions/55991018/whats-the-data-format-of-athenas-csv-metadata-files for more info on the format of these files.

The idea is that it should be quicker to load the metadata file than calling GetQueryResults. Initial investigation reveals that it actually isn't, at least not by very much. I probably have to switch to the async S3 client and load the metadata and rows at the same time.