Use an async S3 client and load the metadata and rows in parallel
This is more of a spike than a great implementation, there is a lot of cases that aren't considered or covered by this. The AsyncRequestTransformer API is really big and complex and there's a lot that can go wrong.

I'm not even sure this is the way to go, I don't know if the performance is good enough and if I'm using the reactive streams API properly (what's the unit of Subscription#request, for example? is it bytes or is it chunks?).