Work around JDK 8/9 issues
Sometimes I get this error when running the tests:

    java.lang.NoSuchMethodError: java.nio.ByteBuffer.flip()Ljava/nio/ByteBuffer

This seems to be due to changes in ByteBuffer between JDK 8 and 9. I've configured Maven and IntelliJ to both compile and run with JDK 8, which should avoid this problem, but I still get it. I assume there is a setting somewhere I have missed. The error should only happen when the code has been compiled with JDK 9 and is run with JDK 8.