Improve the test coverage of S3Results
This adds coverage for some exception handling (including a fix where I had been sloppy and used `CompletableFuture#get()` instead of `#get(long,TimeUnit)`), and defined behaviour for fetch size.

I tried adding coverage for handling InterruptedException from `CompleteableFuture#get(…)`, but I couldn't get it to work. The problem is that the futures that the tests control are not the futures that `#get(…)` is called on, there's chained processing and combinations going on.