Replace ConnectionConfiguration stubs with something more robust
I got really complicated concurrency issues in the tests when I tried to stub ConnectionConfiguration to get it to return the right things at the right time. A separate implementation that can be created from a set of functions works much better.

This changes ConnectionConfiguration to an interface to make it all work.

It wasn't strictly necessary to replace the mock in AthenaResultSetTest, but doing so makes it easier when those tests need to control the configuration.