Make ConnectionConfiguration manage all resources
Instead of AthenaConnection closing just the Athena client, it should close all resources that has been created to support the connection, like the S3 client – but it can't know that an S3 client has been created. This is solved by making the factory and client cache, ClientConfiguration, also be the thing that manages  stateful resources and handles cleanup.