Improve error handling
This makes error handling consistent between the large try blocks in the result classes and the AthenaStatement class by making sure we unwrap ExecutionException in all instances, and that we throw something for all instances of RuntimeError.

In addition, this changes the logic so that we add the outer exception as a suppressed exception in the event that we unwrap an exception to find a cause. From an interface perspective, these wrappers should be internal concepts, but for purposes of debugging, keeping any additional context available in them might be helpful.