Split handling of network and query timeout
The configuration used to only have an API timeout, and if my reading of the JDBC documentation is correct, this would correspond the specified network timeout for JDBC, although for proper handling it should probably also go into the HTTP client inside the AWS SDK in order to ensure we close sockets promptly and not leave operations running.

As far as I understand, the query timeout should instead apply to the full query operation, and not the individual network requests. So far, we retain the old functionality for query timeouts in addition to adding support for network timeouts.