Add a query-timeout based deadline to AthenStatement
This replaces the CompletableFuture.get operations with timeouts corresponding to the time remaining, and propagates the deadline across the polling.

One of the tests in AthenaStatement is now very slow, because you cannot set a non-zero query timeout lower than one second, and it wants to detect the distinction between one operation being delayed and multiple different operations being delayed.