Cancel a query after timeout passed
The purpose of cancelling the query is to avoid consuming all Athena resources on queries that nobody is waiting for anyhow. If the timeout was passed before StartQueryExceution finished, we can't easily cancel it, as we do not yet have a query execution ID.

This doesn't actually wait for the cancellation to go through, and any failure reported in cancellation will just be ignored. Given that we have already used up all the time the user said we had, it seems unreasonable to wait for the cancellation to finish, but there is also no natural way to report an asynchronous error.