WIP: Introduce a module declaration for Java 9+
It turns out to be difficult to do this in a reasonable way, where the sources are still compiled in a such a way that the resulting JAR can be used in Java 8, and the project can be compiled with a Java 8 SDK, and it can be used without problems in IntelliJ.

In order to allow building for Java8, we need to specify release and target to Java8, and instruct the module-info.java to be ignored. This seems to work from the command line, but IntelliJ still seems confused about the module-info.java file.

In order to build for Java9 (in a backwards compatible manner) things become a bit more complicated. First, we need to introduce two compile executions as we cannot compile the module-info.java file using release/source/target 8, but we have to specify that for the remaining files. In principle, it would be nice to actually build using release 9 with target 8, to validate that the module definition includes everything that it has to, but it doesn't seem to be possible as the module definition is ignored if its source file is not included, but not compilable if it is. Thus, we end up effectively building the sources twice, first in the default compile we compile everything targeting Java9, and then in the java8-compile we compile everything except the module-info.java to be compatible with Java8.

Second, in order to actually run the tests, we need to allow reflection on the types in our packages (I think particularly in order to be usable with Mockito).

With all that in place, it seems to be buildable from the command-line using both Java8 and Java9, and it seems to work in IntelliJ as well.

As for the module declaration itself, it is rather extensive, which is primarily caused by the fact that AWS doesn't have proper module declarations. Also, it turns out that AWS hasn't even added the appropriate module name declarations everywhere, so you get a message discouraging publishing the artifact publicly.