Throw exceptions when interrupted
In the presence of an InterruptedException, you should either re-raise it or set the interrupted flag. In our case, we can't really re-raise the exception as is, as it is a checked exception and the throws clauses are constrained by interfaces.

The previous implementation returned something simple, but that result is typically meant to signify something else. For example, returning -1 from a InputStream.read operation signals that we successfully reached end-of-file, whereas an InterruptedException typically means we were aborted before then, and signaling end-of-file would typically result in a parse failure that effectively discards the original exception and reason for failure.

Thus, this instead changes the behavior to wrap the InterruptedException in another Exception with the required signature matching the interface throws clause. In a caller where you once again control the return signature, you might want to check the interrupted flag and throw a new InterruptedException.