Skip empty byte buffers in InputStreamResponseTransformer
We have seen situations where the an IOException with "Underlying input stream returned zero bytes" is thrown. This means that InputStreamResponseTransformer.read has returned zero, which is not allowed. This can occur if an empty byte buffer is published. We have not looked into why, but it happens consistently for certain files.

Since there are no tests for InputStreamResponseTransformer, a test has been added to S3ResultTest instead.