Make tests timing independent
Replace the actual delays in the tests with simulated delays using the test clock.

In order to simulate this, a subclass for delayed completable futures is introduced, that implements the get method by checking whether the explicit delay specified in the constructor is within the deadline specified by the arguments.

Unfortunately, the completable future interface is absolutely massive, and we don't want to implement every method. Luckily, we are (thus far) only using the get method with an explicit time period, and thus it is enough to implement that.

In order for the tests to not become silently meaningless if we start using additional aspects of the interface, the delayed feature is wrapped using a restrictive mock object that raises an exception when any other method is called.