Also make S3Result tests time-independent
This extends the delayed-completable future implementation to actually support thenApply and thenCombine as well.

Also, this changes so that the clock is only ticked once, by executing it along the body in an executor. Technically, this means that the clock might move further than it should when a deadline is applicable, but we do not (yet) have any tests that rely on timing information after a recovery from a timeout exception, so that shouldn't really matter.

This also changes the get logic to assume that the test clock is initialized to zero at the start of each test. Otherwise, we have to solve a complicated synchronization problem to make sure that no ticks happen before the (relevant) get call, or else the deadline will be wrong.