Change timeouts to be higher magnitudes
Now that the delays are not applied to wall time, there is no reason to keep them in millisecond intervals. In fact, the previous tests are still effected by time as the wrapped futures still operate using real time. With the millisecond intervals, it would happen from time to time that the actual execution of the test took longer than the specified timeout (e.g. due to concurrent garbage collection), so this changes the unit to seconds to make this very unlikely.