Delay error handling in InputStreamResponseTransformer
Before this, the input stream would be rendered defunct as soon as an error was generated, but given that we have buffered several megabytes worth of data, it would make more sense to actually push that data through before failing. It might be that input stream consumer doesn't care about the complete response, but only the first portion, and then it seems wasteful to discard the buffered data.