Allow delayed pagination of results
Before this change, the results from S3 are streamed using a single request, which works well when the results are immediately consumed (and works reasonable well when the results are processed in a constant relatively slow rate). But if you process the result set in pages, with large delays between each chunk, it breaks down, as the S3 connection is terminated.

While it would be possible to fix this by keeping the connection to S3 open for longer time periods, that is generally not a good idea, and would in any case be hampered by any transient connectivity issues.

This changes the downloading of the Athena result CSV to handle errors while streaming the results so that another S3 request is restarted at the offset where the previous one failed. This means that a pause between pages will essentially be handled as if we kept the connection open, but with the added benefit that we don't keep the resources allocated, and that we could handle transient network errors in a similar fashion.

This extends the get object helper to allow signaling early exceptions and to remove an object publisher.