Improvements to GetObjectHelper
This improves the helper to correspond more closely to reality in two ways. First, before triggering onError, the code also triggers exceptionOccurred. Second, instead of throwing a NoSuchKeyException immediately, the exception is instead wrapped in a future, to simulate that an actual request is performed.

This also deletes a couple of lines of code that is automatically provided by the base class so that this handles a get-object request in addition to a builder consumer.