Add retries to S3 input stream
Before this change, the results from S3 are streamed using a single request, which works well when the results are immediately consumed (and works reasonable well when the results are processed in a constant relatively slow rate). But if you process the result set in pages, with large delays between each chunk, it breaks down, as the S3 connection is terminated.

While it would be possible to fix this by keeping the connection to S3 open for longer time periods, that is generally not a good idea, and would in any case be hampered by any transient connectivity issues.

This changes the downloading of the Athena result CSV to handle errors when reading the results so that another S3 request is restarted at the offset where the previous one failed. This means that a pause between pages will essentially be handled as if we kept the connection open, but with the added benefit that we don't keep the resources allocated, and that we could handle transient network errors in a similar fashion.

An alternative approach was first attempted that tried to inject similar behavior between the S3 client and the AsyncResponseTransformer, but this interface is rather complex, and is interacted with from (at least) two threads. The current solution is simpler in that it is injected in the single-threaded read-loop.

This produces slightly different results, as a reconnection attempt wouldn't happen until we have consumed the whole input buffer, whereas the previous approach could in theory provide lower latency. However, for the pagination case, it seems more reasonable to wait a bit longer or else we are likely going to practically keep a connection open the entire time.